feat(skills): replace the editor dialog with a create wizard and a per-skill page
